move Footer component to Home page's component folder
